change set_value &add set_glo_list_id

+++ main.py
@@ -35,7 +35,6 @@ class HydraArgParse():
             '-d',
             action="store_true",
             dest="delete",
-            default='',
             help="to confirm delete lun")
         self.parser.add_argument(
             '-s',
@@ -113,7 +112,6 @@ class HydraArgParse():
         User determines whether to delete and execute delete method
         '''
 
-
         crm = vplx.VplxCrm()
         list_of_del_crm = crm.crm_show()
 
@@ -141,8 +139,8 @@ class HydraArgParse():
 
     def execute(self, dict_args):
         for id_one, str_one in dict_args.items():
-            consts.set_value('ID', id_one)
-            consts.set_value('STR', str_one)
+            consts.set_glo_id(id_one)
+            consts.set_glo_str(str_one)
             self.transaction_id = sundry.get_transaction_id()
             self.logger = log.Log(self.transaction_id)
             self.logger.write_to_log(
@@ -152,7 +150,7 @@ class HydraArgParse():
             if self.list_tid:
                 tid = self.list_tid[0]
                 self.list_tid.remove(tid)
-                consts.set_value('TSC_ID', tid)
+                consts.set_glo_tsc_id(tid)
 
             self._storage()
             self._vplx_drbd()
@@ -174,21 +172,15 @@ class HydraArgParse():
         if args.id_range:
             ids = [int(i) for i in args.id_range.split(',')]
 
-
-        if args.delete:
-            if args.uniq_str:
-                consts.set_value('RPL', 'no')
-                if args.id_range:
-                    ids = [int(i) for i in args.id_range.split(',')]
-                else:
-                    ids = ''
-                consts.set_value('STR', args.uniq_str)
-                consts.set_value('ID', ids)
-                self.delete_resource()
+        if args.delete and args.unique_str:
+            consts.set_glo_rpl('no')
+            consts.set_glo_str(args.uniq_str)
+            consts.set_glo_id_list(ids)
+            self.delete_resource()
 
         elif args.uniq_str and args.id_range:
-            consts.set_value('RPL', 'no')
-            consts.set_value('LOG_SWITCH', 'yes')
+            consts.set_glo_rpl('no')
+            consts.set_glo_log_switch('yes')
             if len(ids) == 1:
                 dict_id_str.update({ids[0]: args.uniq_str})
 
@@ -200,8 +192,8 @@ class HydraArgParse():
                 self.parser.print_help()
 
         elif args.replay:
-            consts.set_value('RPL', 'yes')
-            consts.set_value('LOG_SWITCH', 'no')
+            consts.set_glo_rpl('yes')
+            consts.set_glo_log_switch('no')
             logdb.prepare_db()
             db = consts.glo_db()
             if args.transactionid:
@@ -211,7 +203,7 @@ class HydraArgParse():
                     print(
                         f'事务:{args.transactionid} 不满足replay条件，所执行的命令为：python3 {cmd}')
                     return
-                consts.set_value('TSC_ID', args.transactionid)
+                consts.set_glo_tsc_id(args.transactionid)
                 dict_id_str.update({id: string})
 
                 # self.replay_execute(args.transactionid)
